WIP tweaks to StoreStatsViewController
